MB-22960 - Ensure full backfill when cursor is dropped
Due to cursor dropping we re-register the replication cursor only
during backfill when we mark the disk snapshot.  However by this point
it is possible that the persistence cursor has moved ahead.  This would
mean we would lose data when we switch over to in-memory streaming from
backfilling.

To avoid this data loss when we register the cursor we check to see if
the result minus one is greater than the requested starting seqno. If
so we know we have missed some items and we must perform another
backfill.

The reason for the minus one is that registerCursorBySeqno returns the
bySeqno with which the cursor can start and therefore we want to ensure
the previous end is not greater than the requested starting seqno.

Change-Id: Ie769ad0210e7dc3b435a1f9ccfeb4114b995e485