Refactor: CacheCallback to not directly access HashTable object Ver. 2
The hash table object should ideally be a protected member of the
VBucket class, as opposed to being public.

The patch moves towards this goal by introducing a getItem function to
the VBucket class, which ensures that the backfill function
CacheCallback::callback no longer requires direct access to the hash
table object.

Tests to ensure the refactor does not change functionality were added
in the following patch http://review.couchbase.org/#/c/78650/

Note: This patch performs the same refactoring to one that has already
been merged.  However it was found to have a bug and therefore was
reverted.
For original refactor see:
 - Change-Id: I2c71dd30f6b8c66de668137cac1bf358bbdf6588
For the revert of the refactor see:
 - Change-Id: I5dc4e7dad4bb4a91ff94eb7509499ec8f62874b5

In addition to addressing the bug found in the original patch, this
version has an additional regression test to protect from the bug being
re-introduced.

Change-Id: I67f2693d8cbabb30be5b9cc72d550b8f68ee6316