MB-28047: Change the update interval to be percent of items in HT
The update of the frequency counter threshold is currently performed
every time 100 items have been visited in the hash table.

After performing an experiment on MacPro Intel i7-2.5GHz it was shown
that for 100 million items in the hash table it takes a total of 0.5
seconds to calculate the frequency threshold using the method
ItemEviction:getFreqThreshold.

Therefore we want to reduce how often the function is called. So this
patch changes from an absolute value (of 100 items visited) to be a
percentage of the number of items in the hash table.

After experimentation it had been decided to set the interval at 0.1%
of the number of items in the hash table.  With the exception that the
interval must not be below 100 items.

Note: We have a learning phase for the first 100 items visited, which
means we recalculate the frequency threshold for up to the first 100
items.

Change-Id: I03239d87c0c3d137dea31dd2a6bdedded8acdc14