MB-28047: Update the memory recover target after visiting each vbucket
With the StatisticalCounter eviction policy we moved away from the
original policy of recalculating the memory target for every vbucket
that we visit.  Instead we calculated the memory target on the
invocation of the ItemPager::run method and kept it constant whilst
iterating over all vbuckets.

However this has been found to cause problems, particularly when there
is a constant flow of new items being stored in the hash table.
Therefore this patch reverts back to recalculating the memory target
when it visits each vbucket.

As we are recalculating the memory target we can remove the need for
the evictionPercent and the associated evictionMultipler.

Change-Id: I6f25a90a4da9a382fb022fc7924c3c1568ba255f