MB-28047: Remove unrequired checkspoints immediately after eviction
Currently in the PagingVisitor we remove checkpoints that are both
closed and referenced before calculating how much memory to recover, and
then visiting each items to see if it can be evicted.

However we do not then attempt to remove any newly unrequired
checkpoints after the items have been evicted.  We simply move onto the
next vbucket.

This means that we potentially evict more items than we need to.  For
example when using the 2-bit_lru eviction policy on the
STHashTableEvictionItemPagerTest it evicts 91569 items.  Howver if
we repeat the checkpoint removal immediately after eviction, before
moving onto the next vbucket, the number of items required to be
evicted reduces to 91343.

This patch moves the removal code into a private function and calls it
both at the start of PagingVisitor:visitBucket and immediately after
performing eviction for a vbucket.

Change-Id: I744480aa0f0b49fab001113e70f96599f63c291f