MB-29013: Increase initialFrequencyCount from 5 to 64
The new hifi_mfu eviction policy does not perform as well as the
original 2-bit_lru eviction policy when running the YCSB workload D.

This workload has a temporal access pattern (see
https://github.com/brianfrankcooper/YCSB/wiki/Core-Workloads) and
therefore the issue is that the policy is not keeping new items
in the hash table.

The initialFrequencyCount is designed to help ensure that new items are
not immediately evicted from the hash table.  The value was set to 5
which corresponds to (5/65000)*100 = ~0% of the total number of
'increments' before the counter becomes saturated.  It has become clear
that this value is too low.

We want to increase this to approximately 5%, as this will help to avoid
new items from being immediately evicted.  To achieve the 5% we want to
have 'incremented' the counter approximately 3250 times - because
(3250/65000)*100 = 5%.  This corresponds to a probabilistic counter
value of approximately 64.

Change-Id: I05fff400affdc58cac31d24e706f0c3d80d8fa7e