MB-29013: Add temporal behaviour to hifi_mfu eviction policy
The YSCB workload D is a highly temporal workload.  The current
hfi_mfu eviction policy only relies on execution frequency to make
eviction decisions.

We are therefore are introducing a temporal element into the eviction
decision.  The rule is that if an item has an execution frequency
below the freq_counter_age_threshold (defaulted to 1) then its
eviction decision will be based solely on being below the execution
frequency (and if its eligible for eviction).

If an item has an execution counter greater than or equal to
freq_counter_age_threshold then we also take age into account.  We only
evict if the item frequency count is less than or equal to the frequency
threshold and the age is greater than or equal to the ageThreshold.

The ageThreshold is generated using an approach similar to the frequency
threshold.  We maintain an ageHistogram, and set the ageThreshold from
reading out a value from the histogram at a given percentage.  The
percentage is a configuration parameter (defaulted to 20%).

Change-Id: I21378d6a8260c4e9d8908712fcbd4d763794c054