MB-32591: Re-Introduce ChkPtQueueIterator
To help reduce memory usage we want to be able to remove items from the
checkpoint queue where all cursors are to the right (i.e. the items have
been processed) and hence don't need to be held in the queue.

The first step in this is to wrap the checkpoint queue iterator, (an
iterator on a stl::list) in a special iterator that automatically
skips passed elements that contain null pointers.

Change-Id: I88fdb500567f6ec03bf14c6bbbff794a79888373