MB-33423: Temporarily disable CheckpointIterator exception
CheckpointIterator contains exception code to cover the unexpected
case of attempting to access from the end() position.  As this is
invalid an exception is thrown.

This exception was triggered during system test however the undelying
cause has yet to be established.  Therefore temporarily disabling
the exception so that if the issue occurs again a mini-dump will be
generated which can be used to identify the underlying issue.

It is intended that the patch will be reverted prior to the final
release of Mad-Hatter.

Change-Id: I6eb0f00e6c900a11946d414ef04e86add97e0019